oglGameVars now loads and populates levelList, initializes player variables and I added getFilename and getDesc functions to support this move. The .h file also has the struct listed about the class to support the move.
playGame levelList has been removed since it is now in oglGameVars. Because of this I changed the way the levels are loaded. 

Succeed state is now deleted and replaced with SaveGameState (although this may be deleted as well in the near future)

LoadGameState now accepts user input and will load the file name typed in if it exists (still need code to prevent it from attempting to load a file that doesnt exist)

changed levelviewstate to levelselectstate. 

4th button of the main menu now goes to levelselectstate instead of options for my debugging purposes. 

StartNewGameState now prompts user for input, sets all values and goes right into the game after gathering user input instead of displaying a menu as it did before. 

--that's all for this update, tomorrow's will have the menus all cleaned up with more error checking and functionality added.